Updated primitive shapes, added new collision function
Added an Area3D and collisionShape3D duo to each of the primitive shapes, these will be used for snapping technologies later on, where when the area3d collides with an object it will find the closest plane to mesh and combine to.

For the collision function, when the user picks up an object it will disable its physical collision so it does not collide with objects while the user is moving it, and upon release it will return its collision properties.