Fixed remove state Script, added new scripts
When changing tabs using UI Controller to remove, signal from UIC script sends the index value for the tab to the remove script, if the value of the script is 1 then it will be able to run the function commands in the script.
Also added the remove script specifically, moving to a more state machine like controlling for the user.