Update gitignore, updated remove script and summon script
Changes in summon script:

Summon script can no longer remove objects from the scene.
Using signalling similar to remove script, the summon script will not function unless the user is on the build tab, it now highlights as brown to see if you are on build mode.

Has a new signal to emit to other tool scripts when a new object has been summoned to the scene, this is called in the summon_object function.
use set_page_index from the remove script.

Remove script changes:

connect the summon controller in the code, to check if the objectSummoned signal has been emitted, if it is true, update the current summonedArray in the script.

Use existing functions from the Summon script here as well.

Gitignore changes:

added .tmp to gitignore files to ignore any temporary files.