Remove signal added to Summon script
Remove changes:

Added new remove signal that emits a signal to anything connected to the script similar to how the summon script works.

summoner changes:

Summoner script connects to the remove script and gets the signal when an object is removed from the scene by remove script, then it updates its list.