Combine function created
Summon Script Changes:
Combine_objects function now exists, this function works as follows, when the user is placing an object down, a raycast will be checking to see if it collides with any objects, if it does it will place the new object onto its collision point, as well as reparent that new object to the object it is now attached too.

Highlighting here does not fully work however, it only highlights the parent object.

Removing function has not change it still removes the object from the scene, however you remove the entire object when trying to remove one child.