Changes in Summon script and remove script
Remove Script Changes:

onready the summonedObjects array now gets updated when the software starts.

Summon Script Changes:

combine_objects now takes in 4 parameters, the previous 2 summonIndex and obj, and now I split the raycast point and raycast normal.

Now in the function the transform origin is done manually with the formula and the look at function now uses the spawnPoint and the object normal to make the object face the plane it is hitting.