Edit Script updates
Edit Script changes:

Added in variable to hold the objects current position before the change.
Added in the is_active variable that only allows this to work when the edit tab is open on the UI Controller.
Only allow the grip to activate while an object is highlighted, meaning the user is hovering over an object.